+++ app/src/main/java/com/example/awaq_agromo/components/dashboardcomponents/MisCultivos.kt
@@ -1,0 +1,75 @@
+package com.example.awaq_agromo.components.dashboardcomponents
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import com.example.awaq_agromo.presentation.screens.dashboard.CropItem
+import com.example.awaq_agromo.ui.theme.PrincipalSecondary
+
+
+@Composable
+fun MyCropsSection(crops: List<CropItem>) {
+    Column {
+        Text(
+            text = "Mis cultivos",
+            style = MaterialTheme.typography.titleLarge,
+            fontWeight = FontWeight.Bold,
+            color = MaterialTheme.colorScheme.onSurface
+        )
+        Text(
+            text = "Acceda y gestione tus cultivos",
+            style = MaterialTheme.typography.bodyMedium,
+            color = MaterialTheme.colorScheme.onSurfaceVariant
+        )
+        Spacer(Modifier.height(16.dp))
+        LazyRow(
+            modifier = Modifier.fillMaxWidth(),
+            horizontalArrangement = Arrangement.spacedBy(10.dp)
+        ) {
+            items(crops) { crop ->
+                CropIcon(crop.iconRes, crop.description)
+            }
+        }
+    }
+}
+
+@Composable
+fun CropIcon(iconRes: Int, description: String) {
+    Column(horizontalAlignment = Alignment.CenterHorizontally) {
+        Box(
+            modifier = Modifier
+                .size(64.dp)
+                .clip(CircleShape)
+                .background(MaterialTheme.colorScheme.surfaceVariant)
+                .border(width = 4.dp, color = PrincipalSecondary, shape = CircleShape),
+            contentAlignment = Alignment.Center
+        ) {
+            Image(
+                painter = painterResource(id = iconRes),
+                contentDescription = description,
+                modifier = Modifier.size(42.dp)
+            )
+        }
+        Spacer(Modifier.height(4.dp))
+        Text(text = description, style = MaterialTheme.typography.labelSmall)
+    }
+}
